Added new form to use in BookPreview validation.

+++ books/views.py
@@ -1,10 +1,11 @@
 from django.views.generic import CreateView, FormView, TemplateView
 from django.contrib.auth.mixins import LoginRequiredMixin
 from django.urls import reverse_lazy
-from django.http import HttpResponseForbidden, HttpResponseRedirect, Http404
+from django.http import HttpResponseForbidden, HttpResponseRedirect
+from django.shortcuts import get_object_or_404
 
-from .models import BookCopy, Book, BookCoverPreview
-from .forms import BookForm
+from .models import BookCopy, BookCoverPreview
+from .forms import BookForm, BookPreviewForm
 
 
 class BookCopyCreateView(LoginRequiredMixin, CreateView):
@@ -26,7 +27,7 @@ class BookCreateView(LoginRequiredMixin, FormView):
 
     def form_valid(self, form):
         user_profile = self.request.user.userprofile
-        BookCoverPreview.objects.filter(profile=user_profile).delete()
+        BookCoverPreview.objects.get_object_or_404(profile=user_profile).delete()
         cover_preview = BookCoverPreview(profile=user_profile, cover=form.cleaned_data.pop('cover'))
         cover_preview.save()
         self.request.session['book'] = form.cleaned_data
@@ -48,18 +49,15 @@ class BookPreviewView(LoginRequiredMixin, TemplateView):
         if request.method == 'POST':
             if 'book' in self.request.session:
                 book_data = self.request.session.pop('book')
-                book = Book()
-                book.title = book_data['title']
-                book.author = book_data['author']
-                book.description = book_data['description']
-                book.page_count = book_data['page_count']
-                try:
-                    cover_preview = BookCoverPreview.objects.get(profile=request.user.userprofile)
-                except BookCoverPreview.DoesNotExist:
-                    return Http404()
-                book.cover = cover_preview.cover
-                book.save()
-                cover_preview.delete()
+                cover_preview = get_object_or_404(BookCoverPreview, profile=request.user.userprofile)
+                form = BookPreviewForm(book_data)
+                if form.is_valid():
+                    book = form.save()
+                    book.cover = cover_preview.cover
+                    book.save()
+                    cover_preview.delete()
+                else:
+                    return HttpResponseForbidden()
                 return HttpResponseRedirect(reverse_lazy('library_list'))
             else:
                 return HttpResponseForbidden()
